Refactor: Move fxhash_32 to its own file
